recover complete

+++ squirrel-web/src/main/java/com/dianping/cache/alarm/entity/ScanDetail.java
@@ -15,9 +15,9 @@ public class ScanDetail {
 
     private String rdDuty;
 
-    private int totalCount;
+    private long totalCount;
 
-    private int failCount;
+    private long failCount;
 
     private double failPercent;
 
@@ -126,20 +126,20 @@ public class ScanDetail {
         return this;
     }
 
-    public int getTotalCount() {
+    public long getTotalCount() {
         return totalCount;
     }
 
-    public ScanDetail setTotalCount(int totalCount) {
+    public ScanDetail setTotalCount(long totalCount) {
         this.totalCount = totalCount;
         return this;
     }
 
-    public int getFailCount() {
+    public long getFailCount() {
         return failCount;
     }
 
-    public ScanDetail setFailCount(int failCount) {
+    public ScanDetail setFailCount(long failCount) {
         this.failCount = failCount;
         return this;
     }
